revert...
Adding lo/hi doesn't help much since this doesn't allow us to magically use the actual scijs/ndarray machinery. Better way may be to wrap our NdArray as a scijs/ndarray in the client code (if it's not already one). This means mapping to a 1D get() since this is what ndarray expects.